Dynamic command deployment & guarding
Commands now get set every time the bot starts. No separate deployment scripts are needed anymore.
Commands are also only set per guild, not globally. For this, a guard function has been added to commands that can check if a command should be added to a guild or not (by e.g. checking if all the necessary things are configured for this guild). This guard is also checked before executing a command.