Make guilds configurable via commands
Instead of predefining guild configs in config file, store the configuration in a database and let server admins change it via commands.
All guild commands get updated on configuration changes, when a guild is joined or manually triggered via a refresh command (as well as still during startup of the bot).