Don't check mentions in /names reply for delete command.
Since the /names reply also mentions an arbitrary user in the last line of it's embed's description we have to make sure that embed isn't checked for mentions in the delete command.
Also combined two ifs and fixed the position of an optional chaining operator in the check for embeds.