Store message metadata in db.
Track information about certain messages sent by the bot in the database.
Mainly used to keep track of the user that triggered the message to be sent when it's not a direct reply to an interaction.
Cleaned up when messages are deleted or after 6 months.
This new metadata is used to better implement the delete command. Bookmarks and quotes don't print the information about the creator anymore. Instead, a new blame context menu command on messages can provide this information.