Upgrade to new permission system.
Instead of searching through the roles of a guild, finding ones that match a permission we want for the bot and setting that role in a command permission, we're making use of the new default member permissions field on commands now.
These will be effective for the command as long as no override has been configured for the command in the server settings.
Along with this, the commands that were restricted to administrators before can now be used by users that have the "manage guild" permission.

For this we needed to update DiscordJS which required us to switch from the deprecated MessageEmbed#addField to MessageEmbed#addFields and to remove an import of memberNicknameMention which we weren't actually using anyway.

I also ensured Prettier doesn't mess with the formatting of the Node package files.