Made commands translateable
Using the name_localizations and description_localizations features added to the Discord API we can now provide translations for command and option names and descriptions and choice names.
I also removed the default descriptions from the command configurations and moved them to the English translation file to avoid duplication.

Because of a show-stopper bug in the Discord desktop client, command names will not be translated for now; otherwise users can not find them by typing the names in the locale they see them in.

Spanish translation will be added later.