Fix translator fetching for commands
Instead of getting the command name from the interaction, we get it from the configuration of the command in the registry. This ensures that interactions received for components (button clicks etc.), which don't have a command name on the interaction, still get the right translation key prefixes in the translator.