Remove IMAGE tag in favour of URL detection
Instead of having to use a special tag, the author of a story can simply put an (escaped) URL in their text and have it show up as an image automatically, since this is how Discord normally works anyway.
To make things a bit easier for authors, lines including URLs are not bundled up with other lines into messages so that the image shows up in the line where the author put the URL, rather than at the bottom of a long block of bundled up lines. That is: lines including URLs are treated the same as lines using the STANDALONE tag.