Players informed about story reset get start button
When uploading a new story file, the story will now be stopped entirely for all players (instead of just resetting the state) and in the notification they receive, they now have a button for starting to play the story again.
To avoid a circular dependency, the function for creating the id for that button had to be imported by config-story.js and passed on to story-engine.js. The button now also contains the guildId (needed when starting the story for verifying getStory). Since this button is not handled in a guild interaction, we don't have the guildId available in the receiving interaction.

Also some small fixes: removed allowed_mentions where it didn't make sense (copy&paste error) and passed through "publicly" parameter in postStory as it was intended.