/story only available if stories exist in server
The /story command is now guarded and will only show up on a server if non-draft stories exist.
This is not ideal as at the moment players can only play published stories. But it's in preparation of implementing test plays which can be done by the editor of a story in status testing, so the command needs to be available then.