Improved story deleting
Removed slash command for deleting story as it's awkward UI and rarely needed, so just cluttering up the list. A button on the show command reply works much better.

Also improved how the command works. It will delete stories immediately if they are in status Testing because it's less of a problem then.

And fixed story deletion in general since for some reason the "ON DELETE CASCADE" clause on plays vanished.