Easier selection of stories in show commands
When using "/story show" or "/config-story show" and there are 25 or fewer stories available, they can now be found in a select menu below the interaction response. Selecting a story from the menu shows the details view of it. This way the user doesn't have to type in the show command again.