Robustness fixes for sending DMs
Handle cases where the user cannot receive DMs from the bot: We try sending one first (before replying to an interaction) and if it doesn't work, we tell the user what to do (and potentially reset/clear some state). We don't log that error since it's not an error on our end but just a user setting.

If this happens when trying to inform the story editor about an issue with the story, there's not much we can do other than ignoring the error.
(Also made a small fix to send multiple messages to editor in sequence instead of in parallel.)

For informing players about their story state being reset, we also ignore the error (about not being able to send them DMs). Additionally, we now handle the case that Discord might tell us we're opening DMs too fast here, and just stop informing the players then. For this I had to separate the play clearing from the informing and also run it in sequence for all the players instead of in parallel.