Removed unused parameter
No caller ever passed the original message. We could get the content / embeds from the message in the interaction, but it seems for updating the message, we only need to pass the parts we want to change, not the whole message.
Maybe it's different in the editReply case but then it was already broken before and will become apparent at a later point.