Use command mentions in bot replies
Instead of just writing the command names as code syntax, we can use the new command mention syntax whenever we talk about a command in a bot reply. That way, the user can actually click on it to use it.
For this, the ids of the commands in each guild have to be runtime-cached.

In one reply where a command was mentioned with a parameter, I changed the translation to get rid of the command mention instead to make it easier. The Spanish translation uses a translated parameter name even though they are not translated for Spanish yet, so it will be wrong for a while.

Unfortunately the mentioned commands will not show up with their localised names (see https://github.com/discord/discord-api-docs/issues/5518 ) and we can't use localised names in the mention either because then they become unclickable.