Use errorReply and warningReply consistently
These two functions that were developed while working on the stories feature are now used for the rest of the commands as well to get a consistent feedback look.

The general guideline when to use which is:
If it's an internal error in the server and not the user's fault / something we'd log as an error, use errorReply. If it could roughly be seen as a user error or telling the user about something that won't work as they probably intended, use warningReply. If it's purely informational or a regular response or success message by the bot, use t.privateReply.
We might introduce more functions like infoReply or successReply in the future.