More reliable guild command updating
Guild commands already get updated when we join a guild or when configuration or any other thing changes that affects the availability of commands in a guild.
Currently, they also get updated for all guilds on restart (but if we stop doing that, we would be able to handle that for guilds the bot joined when it wasn't running).
If updating guild commands on config changes fails, they will now be marked as not updated.
A nightly maintenance job now checks for all guilds we know about where the commands are not updated and attempts to do that.
With this behaviour, things should eventually fix themselves and we don't need /refresh-commands anymore.

If the config was changed in the db but there's an exception before we get to the call to update the commands, then things might be broken.
But the admin can fix that by changing a configuration option again. There's no need for a separate command as it shouldn't happen very often.