Statically import commands and event handlers
Commands and event handlers have been moved around a bit more. Event handlers now have a registry as well.

All of them get statically imported instead of dynamically. This makes the switch to TypeScript easier as we don't have to import from the src directory but can transpile the code.

Not importing things dynamically meant that there were now some circular dependencies to resolve.
The command id cache had to be split into a separate module. And fetching translations for values in a command's configuration had to be deferred. So now the configuration of commands can be generated when they are loaded instead of when their module is loaded.