Convert code to TypeScript
A few small refactorings were necessary to make this work. Some bugs were found and fixed along the way.

Command configurations now require a description property (can be empty) to make them compatible with Discord.JS types. They will still be overwritten by descriptions from the English translation file.

A bug in the typings of inkjs had to be fixed in my fork.
Also the typings of inkjs are not ideal in general. For example it seems I can't import the typings of Choice that are actually provided. And I have to do deep imports of JS classes.

Both inkjs and axios have a problem with their exports and module setup forcing me to access a "default" property on the imported default export.

Tests have not been converted to TypeScript yet.
ESLint has not been configured for TypeScript yet.
In the future, I intend to activate some stricter type checks (like checking for nulls).