Update i18next libs
The version of i18next-fs-backend I was using didn't provide any typings yet. The newer version is based on a new i18next version so I updated that as well.

This triggered a TypeScript error which I couldn't find a workaround for. Apparently people using the React plugin often run into this but it hasn't been reported for the plain library yet. I ignored it with a comment.

This new i18next version extends TS support a lot and apparently provides a way to type the translation resources. But looking at the documentation at https://www.i18next.com/overview/typescript it doesn't look like it would work with resources loaded via .json5 files so I didn't bother integrating it.