Smarter command deployment
Command updating does not happen on every regular startup anymore. Instead, global commands and guild commands can be updated as needed using separate scripts. This should be done whenever the configuration of a command changes, the set of global/guild commands changes or conditions in command guards get changed.
This will speed up startup significantly once the bot is in lots of guilds.

For running these scripts we need the client id in the bot configuration again, since we don't fully connect as a regular client but just make REST requests.
This also means that we don't always have the client object (or the guild object) available in code for updating commands or in guards.
And we have to manually transform command configurations to Discord API data.

The role of updatedGuildsCache changed: Before it was set to true for a guild after commands were updated. Now it is assumed on app start that all commands are regularly updated, and it's set to false only when updating fails for a guild (for example after config changes). So the maintenance job for ensuring guild commands are updated will only act on those.