Fix: Deal with stricter Discord API checks
Discord now seems to check the value property of a text input in a modal against its min_length restriction. So if min-length is set to 1, it will refuse an empty string for the value field.
It seems odd that it should check the prefilled value against those constraints and not just what the user eventually enters, but it's easy enough to work around anyway.