Fix story title constraints
We were setting the titles of newly created draft stories to ''. And because of a unique constraint, there could only be one draft in a server at a time then. Setting them to null instead helps with the unique constraint which counts null values as different from each other. But null values were not allowed in the title column. Now they are allowed in draft status.