Clean up guild config & commands on channel deletion
We now listen to the event of a channel being deleted. We use this to update the guild configuration and guarded commands if necessary, for example if a role-play channel has been removed.

We might not catch all those events, for example if they happened when the bot was down. For this reason I applied two more fixes:
- When an RP channel gets removed because it was detected as not existing anymore during ensuring webhook correctness, we now also update commands for the guild.
- If bookmark/quote commands get called and the channels are removed, their guards already rejected the interaction. Now the user also gets feedback about the command failing.
This doesn't guarantee that bookmarks/quotes channels get removed from the configuration but at least it's handled gracefully and that's good enough.