Retire /names o7
This was one of the more problematic commands (didn't fulfill its purpose that well and would not have worked well in large communities) that also didn't match the profile of the bot well.

While it also allows us to get rid of the GUILD_MEMBERS privileged intent (which we'd have to request once we get verified - and /names probably wouldn't be an acceptable use-case for it), we will have to eventually put GUILD_MEMBERS back in in order to implement user data clean-up.