Document missing Spanish translations
Added all translation keys and English translations that are missing in the Spanish translations file to it. While this isn't necessary for translation purposes (it will fall back to the English translations automatically anyway), this lets me make translation changes more easily while continuing to document what is missing for Spanish.
The file contains a bunch of comments as a guide what to translate and how.