Update inkjs
Still using a fork; based on inkjs 2.2.0.
inkjs 2.2.1 is released but there's been some more changes on its master branch since then that I don't want to pull in yet. Also need to figure out a better version scheme to express which inkjs release my package is based on.

In the meantime, this release lets us use choice tags (which I intend to use for a nicer syntax for button styles) and dynamic tags which might become interesting for scene tags.