Introduce choice tag for receiving user text input
A new tag on choices allows story authors to ask for text input from the user. When clicking the choice button, the user is presented with a modal dialog in which they can enter one line of text. This text will be stored in a variable defined by the story author and the story will then continue with the pressed choice button.